cluster: fix tiflash uptime
