+++ pkg/cluster/operation/check.go
@@ -520,7 +520,7 @@ func CheckServices(ctx context.Context, e ctxt.Executor, host, service string, d
 		return result
 	}
 
-	active, _, err := GetServiceStatus(ctx, e, service+".service")
+	active, _, _, err := GetServiceStatus(ctx, e, service+".service")
 	if err != nil {
 		result.Err = err
 	}
